repair_agent/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.13.2
    # via -r requirements.in
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
async-timeout
    # via redis
attrs==25.3.0
    # via
    #   jsonschema
    #   openapi-python-client
    #   outcome
    #   referencing
    #   trio
blis==1.3.0
    # via thinc
cachetools==5.5.2
    # via google-auth
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   urllib3
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via readability-lxml
charset-normalizer==3.4.3
    # via
    #   reportlab
    #   requests
click==8.2.1
    # via
    #   -r requirements.in
    #   duckduckgo-search
    #   typer
    #   uvicorn
cloudpathlib==0.21.1
    # via weasel
colorama==0.4.6
    # via -r requirements.in
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.2
    # via matplotlib
cryptography==45.0.6
    # via
    #   pyopenssl
    #   urllib3
cssselect==1.3.0
    # via readability-lxml
cycler==0.12.1
    # via matplotlib
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
distro==1.8.0
    # via
    #   -r requirements.in
    #   openai
docker==7.1.0
    # via -r requirements.in
duckduckgo-search==3.0.2
    # via -r requirements.in
exceptiongroup==1.3.0
    # via
    #   anyio
    #   trio
    #   trio-websocket
fastapi==0.116.1
    # via -r requirements.in
fonttools==4.59.0
    # via matplotlib
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.179.0
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-genai==1.29.0
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   openai
    #   openapi-python-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
    #   urllib3
javalang==0.13.0
    # via -r requirements.in
jinja2==3.1.6
    # via
    #   openapi-python-client
    #   spacy
jiter==0.10.0
    # via openai
jsonschema==4.25.0
    # via -r requirements.in
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
langchain==0.3.27
    # via -r requirements.in
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
lxml==6.0.0
    # via
    #   python-docx
    #   readability-lxml
marisa-trie==1.2.1
    # via language-data
markdown==3.8.2
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.5
    # via
    #   -r requirements.in
    #   venn
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==2.2.6
    # via
    #   blis
    #   contourpy
    #   langchain
    #   matplotlib
    #   pandas
    #   spacy
    #   thinc
openai==1.99.9
    # via -r requirements.in
openapi-python-client==0.25.3
    # via -r requirements.in
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   matplotlib
    #   spacy
    #   thinc
    #   weasel
    #   webdriver-manager
pandas==2.3.1
    # via -r requirements.in
pillow==11.3.0
    # via
    #   matplotlib
    #   reportlab
preshed==3.0.10
    # via
    #   spacy
    #   thinc
prettytable==3.16.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via -r requirements.in
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   google-genai
    #   langchain
    #   openai
    #   openapi-python-client
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyopenssl==25.1.0
    # via urllib3
pyparsing==3.2.3
    # via
    #   httplib2
    #   matplotlib
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   openapi-python-client
    #   pandas
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==0.21.0
    # via
    #   -r requirements.in
    #   webdriver-manager
pytz==2025.2
    # via pandas
pyyaml==6.0
    # via
    #   -r requirements.in
    #   langchain
readability-lxml==0.8.1
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.7.34
    # via tiktoken
reportlab==4.4.3
    # via -r requirements.in
requests==2.32.4
    # via
    #   -r requirements.in
    #   docker
    #   duckduckgo-search
    #   google-api-core
    #   google-genai
    #   langchain
    #   spacy
    #   tiktoken
    #   weasel
    #   webdriver-manager
rich==14.1.0
    # via typer
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruamel-yaml==0.18.14
    # via openapi-python-client
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
ruff==0.12.8
    # via openapi-python-client
selenium==4.1.4
    # via -r requirements.in
shellingham==1.5.4
    # via
    #   openapi-python-client
    #   typer
six==1.17.0
    # via
    #   javalang
    #   python-dateutil
smart-open==7.3.0.post1
    # via weasel
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
spacy==3.8.7
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
sqlalchemy==2.0.43
    # via langchain
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
starlette==0.47.2
    # via fastapi
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
thinc==8.3.6
    # via spacy
tiktoken==0.11.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   openai
    #   spacy
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typer==0.16.0
    # via
    #   openapi-python-client
    #   spacy
    #   weasel
typing-extensions==4.14.1
    # via
    #   anyio
    #   cloudpathlib
    #   exceptiongroup
    #   fastapi
    #   google-genai
    #   openai
    #   openapi-python-client
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   python-docx
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3[secure,socks]==1.26.20
    # via
    #   docker
    #   requests
    #   selenium
urllib3-secure-extra==0.1.0
    # via urllib3
uvicorn==0.35.0
    # via -r requirements.in
venn==0.1.3
    # via -r requirements.in
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
wcwidth==0.2.13
    # via
    #   prettytable
    #   prompt-toolkit
weasel==0.4.1
    # via spacy
webdriver-manager==4.0.2
    # via -r requirements.in
websockets==15.0.1
    # via google-genai
wrapt==1.17.3
    # via smart-open
wsproto==1.2.0
    # via trio-websocket
gitpython==3.1.45

# The following packages are considered to be unsafe in a requirements file:
# setuptools
